Revert prefabs and outpoststation
