+++ UnityProject/Assets/Resources/Prefabs/Chat/ChatEntry.prefab
@@ -82,9 +82,7 @@ MonoBehaviour:
     m_HorizontalOverflow: 0
     m_VerticalOverflow: 1
     m_LineSpacing: 1
-  m_Text: 'test
-
-'
+  m_Text: 
 --- !u!114 &114862419011751618
 MonoBehaviour:
   m_ObjectHideFlags: 0
