+++ UnityProject/Assets/Scripts/Chat/ChatEntryPool.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4f577727280d45f4a91b2a0bd15fbfea
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
